💦 Refactor: ChatRoom, ChatParticipate 패키지 이동
